feat(010): allow floats to be used

+++ day-010/project-010_calculator/main.py
@@ -29,9 +29,9 @@ def main():
     print(art.logo)
 
     calculating = True
-    num1 = int(input("What's the first number?: "))
+    num1 = float(input("What's the first number?: "))
     while calculating:
-        num2 = int(input("What's the second number?: "))
+        num2 = float(input("What's the second number?: "))
         for symbol in operations:
             print(symbol)
         operation_symbol = input("Pick an operation from the options above: ")
